# Conflicts:
#	src/Chemistry/BindingManager.cpp
#	src/Chemistry/ChemCallbacks.h
#	src/Chemistry/HybridBindingSearchManager.cpp
#	src/Controller.cpp
#	src/MathFunctions.h
#	src/Mechanics/CUDAcommon.cpp
#	src/Mechanics/CUDAcommon.h
#	src/Mechanics/ForceField/Boundary/BoundaryBubbleRepulsion.cpp
#	src/Mechanics/ForceField/Boundary/BoundaryCylinderRepulsion.cpp
#	src/Mechanics/ForceField/Boundary/BoundaryCylinderRepulsionIn.cpp
#	src/Mechanics/ForceField/Bubble/BubbleBubbleRepulsionExp.cpp
#	src/Mechanics/ForceField/Bubble/BubbleCylinderRepulsion.cpp
#	src/Mechanics/ForceField/Bubble/BubbleCylinderRepulsionExp.cpp
#	src/Mechanics/ForceField/Bubble/MTOCAttachment.cpp
#	src/Mechanics/ForceField/Filament/FilamentBendingCosine.cpp
#	src/Mechanics/ForceField/ForceFieldManager.cpp
#	src/Mechanics/ForceField/ForceFieldManager.h
#	src/Mechanics/ForceField/MotorGhost/MotorGhostStretching.cpp
#	src/Mechanics/Minimizer/CGFletcherRievesMethod.cpp
#	src/Mechanics/Minimizer/CGMethod.cpp
#	src/Mechanics/Minimizer/CGMethod.h
#	src/Mechanics/Minimizer/CGPolakRibiereMethod.cpp
#	src/Output.cpp
#	src/Output.h
#	src/Restart.h
#	src/Structure/Bead.cpp
#	src/Structure/Bead.h
#	src/Structure/BinGrid.h
#	src/Structure/BoundaryElement.h
#	src/Structure/BoundaryImpl.h
#	src/Structure/Bubble.h
#	src/Structure/CMotorGhost.cpp
#	src/Structure/Cylinder.cpp
#	src/Structure/Cylinder.h
#	src/Structure/Filament.cpp
#	src/Structure/Filament.h
#	src/Structure/HybridNeighborListImpl.cpp
#	src/Structure/Linker.h
#	src/Structure/MotorGhost.cpp
#	src/Structure/MotorGhost.h
#	src/Structure/Special/MTOC.h
#	src/Structure/SubSystem.cpp
#	src/Structure/SubSystem.h
#	src/utility.h

+++ src/utility.h
@@ -35,7 +35,7 @@
 using namespace std;
 
 ///floatingpoint typedef
-typedef double floatingpoint;
+typedef float floatingpoint;
 typedef double doubleprecision;
 
 namespace detail {
